refactor(amp): create amp dir | move helpers in individual files
