+++ amp/helpers/index.js
@@ -1,0 +1,5 @@
+const { InfiniteScrollAmp } = require("./infinite-scroll");
+const { setCorsHeaders } = require("./set-cors-headers");
+const { optimize } = require("./optimize-amp-html");
+
+module.exports = { InfiniteScrollAmp, setCorsHeaders, optimize };
