+++ amp/helpers/optimize-amp-html.js
@@ -1,0 +1,13 @@
+const AmpOptimizer = require("@ampproject/toolbox-optimizer");
+
+const ampOptimizer = AmpOptimizer.create({
+  autoAddMandatoryTags: false,
+  autoExtensionImport: false,
+  preloadHeroImage: false,
+});
+
+async function optimize(ampHtml) {
+  return ampOptimizer.transformHtml(ampHtml);
+}
+
+module.exports = { optimize };
