+++ server/routes.js
@@ -131,14 +131,14 @@ function withConfigPartial(
   getClient,
   logError,
   publisherConfig = require("./publisher-config"),
-  configWrapper = config => config
+  configWrapper = (config) => config
 ) {
   return function withConfig(f, staticParams) {
     return function (req, res, next) {
       const client = getClient(req.hostname);
       return client
         .getConfig()
-        .then(config => configWrapper(config))
+        .then((config) => configWrapper(config))
         .then((config) =>
           f(
             req,
@@ -285,8 +285,8 @@ exports.isomorphicRoutes = function isomorphicRoutes(
     cdnProvider = "cloudflare",
     serviceWorkerPaths = ["/service-worker.js"],
     maxConfigVersion = (config) =>
-    get(config, ["theme-attributes", "cache-burst"], 0),
-    configWrapper = config => config,
+      get(config, ["theme-attributes", "cache-burst"], 0),
+    configWrapper = (config) => config,
 
     // The below are primarily for testing
     logError = require("./logger").error,
@@ -296,8 +296,12 @@ exports.isomorphicRoutes = function isomorphicRoutes(
     publisherConfig = require("./publisher-config"),
   }
 ) {
-
-  const withConfig = withConfigPartial(getClient, logError, publisherConfig, configWrapper);
+  const withConfig = withConfigPartial(
+    getClient,
+    logError,
+    publisherConfig,
+    configWrapper
+  );
 
   pickComponent = makePickComponentSync(pickComponent);
   loadData = wrapLoadDataWithMultiDomain(publisherConfig, loadData, 2);
@@ -545,7 +549,8 @@ exports.mountQuintypeAt = function (app, mountAt) {
 /**
  * *ampRoutes* handles all the amp page routes using the *[@quintype/amp](https://developers.quintype.com/quintype-node-amp)* library
  * routes matched:
- * *"/amp/story/:slug"* returns amp story page.
+ * GET - "/amp/story/:slug"* returns amp story page
+ * GET - "/amp/api/v1/amp-infinite-scroll" returns the infinite scroll config JSON. Passed to <amp-next-page> component's `src` attribute
  *
  * @param {Express} app Express app to add the routes to
  * @param {Object} opts Options object used to configure amp. Passing this is optional
